changes to allow maketarget work with makeclothes
Scale and basemesh are needed from external sources. If a standard
basemesh is loaded, scale should be 1.0. So if the attribute isn't there, 1.0 is assumed.

Furthermore I changed output to the one makehuman uses for the standard
targets. Now it is possible to load an existing
standard target, change it and save it, the difference will only be the
new lines.