changes to work with flexible target name
I added a box, where the user can enter the primary target name. This
name will be used as a filename (replacing all non-alpha and non-number
characters to "-"). If a target is loaded, filename will be used as a
target (but linked to object, not to scene). I changed the code to
produce lines a bit, unfortunately I created "0 0 0" lines again.