changes to accept different meshes
maketarget had problems to create a target, when only the body was
loaded. But especially for e.g. hands, face etc. we do not need the
helper. To avoid this I added the mirror table which also tells the
program if a vertex is left, mid or right side.

The table is loaded according to the mesh type, so another table could
be added for other meshes. No hard coding.

Mesh defaults to hm08. I tried with different
meshes, it works. Maybe the data path should be common in the end for
all our tools.

Now maketarget process the mesh and does a lookup in the table for each
vertex. It keeps one side untouched. The vertices in the middle are set to x=0

Before using mirror function object mode is pre-selected.